+++ Cargo.toml
@@ -24,7 +24,6 @@ thiserror = "2.0.16"
 tokio = { version = "1.50.0", features = ["rt", "rt-multi-thread", "fs", "io-util", "sync", "time", "macros"] }
 tracing = "0.1.44"
 tracing-subscriber = { version = "0.3.23", features = ["env-filter"] }
-trait-variant = "0.1.2"
 url = "2.5.8"
 xxhash-rust = { version = "0.8.15", features = ["xxh64"] }
 zip = { version = "8.3.0", features = ["deflate"] }
